requirements.txt
# Core dependencies
playwright>=1.40.0
requests>=2.31.0
lxml>=5.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
py2neo>=2021.2.3

# CLI
click>=8.0.0
pyyaml>=6.0

# Backend
fastapi>=0.100.0
uvicorn>=0.20.0
supabase>=2.0.0
